+++ inbox/crispin.py
@@ -322,6 +322,42 @@ retry_crispin = functools.partial(
 )
 
 
+common_uid_list_format = re.compile(b"^[0-9 ]+$")
+uid_format = re.compile(b"[0-9]+")
+unoptimized_parse_message_list = imapclient.response_parser.parse_message_list
+
+
+def optimized_parse_message_list(data: List[bytes]) -> List[int]:
+    """Optimized version of imapclient.response_parser.parse_message_list
+
+       This takes care of parsing the most common textual format that list may
+       arrive in from an IMAP server. The algorightm is shorter and much less memory
+       hungry than the generic version which becomes important when syncing
+       large mailboxes. It relies on regexes to avoid creating intermediate lists.
+       If we receive data in format that does not follow most common format
+       we still fallback to the unoptimized version.
+
+       Based off Tom's (Python 2 version):
+       https://github.com/closeio/imapclient/commit/475d02f85be308fb4ac80e66628a03c30c096c9f
+       For generic version see:
+       https://github.com/mjs/imapclient/blob/master/imapclient/response_parser.py#L39-L79
+       Implemented in:
+       https://github.com/closeio/sync-engine/pull/483
+    """
+    with contextlib.suppress(TypeError):
+        if len(data) == 1 and common_uid_list_format.match(data[0]):
+            return [
+                int(uid_match.group(0))
+                for uid_match in re.finditer(uid_format, data[0])
+            ]
+
+    return unoptimized_parse_message_list(data)
+
+
+# Replace the unoptimized algorithm with optimized algoritm from above
+imapclient.response_parser.parse_message_list = optimized_parse_message_list
+
+
 class CrispinClient:
     """
     Generic IMAP client wrapper.
